+++ genomepy/functions.py
@@ -1,6 +1,7 @@
 """Module-level functions."""
 import os
 import re
+from tempfile import mkdtemp
 from typing import Optional
 
 import pyfaidx
@@ -20,7 +21,7 @@ from genomepy.files import (
 from genomepy.genome import Genome
 from genomepy.online import check_url
 from genomepy.plugins import get_active_plugins
-from genomepy.providers import download_assembly_report, online_providers
+from genomepy.providers import Provider, download_assembly_report, online_providers
 from genomepy.utils import (
     get_genomes_dir,
     get_localname,
@@ -29,6 +30,18 @@ from genomepy.utils import (
     safe,
     try_except_pass,
 )
+
+
+def head_annotations(name: str, n: int = 1):
+    """
+    For UCSC, up to 4 gene annotation styles are available:
+    "ncbiRefSeq", "refGene", "ensGene", "knownGene" (respectively).
+
+    Quickly inspect the metadata of each available annotation.
+    """
+    tmp_dir = mkdtemp()
+    u = Provider.create("ucsc")
+    u.head_annotations(name, genomes_dir=tmp_dir, n=n)
 
 
 def list_available_genomes(provider=None):